Arg Passing with child/parent kernel threads
Attempt at implementing arg passing using what was discussed in class. There are child and parent kernel threads and a list of child processes. The child kernel thread lives as a separate struct with a tid, shared  list elem, an int to hold exit status that is updated based on successful or unsuccessful execution, and a boolean to know if the thread was used or not.  A semaphore lock is used to ensure that the child dies before the parent executes and it lives within the thread struct.  Code compiles, but currently is not passing any tests.